questions controller voted
