+++ app/controllers/questions_controller.rb
@@ -1,29 +1,9 @@
 class QuestionsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
-  before_action :load_question, only: [:show, :update, :destroy, :vote, :cancel_vote]
+  before_action :load_question, only: [:show, :update, :destroy]
   before_action :check_user, only: [:update, :destroy]
-  before_action :check_voted, only: [:vote]
-  before_action :check_author, only: [:vote]
 
-  def vote
-    if params[:points].to_i < 0
-      @vote = @question.vote_down(current_user)
-      @message = "voted down"
-    else
-      @vote = @question.vote_up(current_user)
-      @message = "voted up"
-    end
-
-    render json: { id: @question.id, message: @message, output: render_to_string(partial: 'votes/block', locals: { data: @question }) }
-  end
-
-  def cancel_vote
-    @vote = @question.votes.where(user: current_user).first
-    @vote.destroy
-    @message = "vote is cancelled"
-
-    render json: { id: @question.id, message: @message, output: render_to_string(partial: 'votes/block', locals: { data: @question }) }
-  end
+  include Voted
 
   def index
     @questions = Question.all
@@ -63,18 +43,6 @@ class QuestionsController < ApplicationController
   end
 
   private
-
-  def check_author
-    if current_user.author_of?(@question)
-      head(:forbidden)
-    end
-  end
-
-  def check_voted
-    if current_user.voted_for?(@question)
-      head(:forbidden)
-    end
-  end
 
   def check_user
     unless current_user.author_of?(@question)
